Rename database configuration folder
